Docs and Ring tweaks
Aside from updating and adding docs, we bid farewell to numbered rings.

So the gen logic can actually check enough rings are reachable...
We need the minimum number of rings to be flagged as progression...
Thus we want to set the progression/filler split at gen...
So we can't just generate the big list as a constant anymore.

We also want to have a cleaner way of checking them, and the numbered
ones pollute the item lists... I liked them individually numbered like
they're some collectable series, but realistically this is saner.
One item that's progression, one item that's filler, multiple copies of
each.